master  copy sync

+++ src/test/java/com/web/BaseWebTest.java
@@ -1,11 +1,12 @@
 package com.web;
 
 import com.tricentis.common.constants.FrameworkConstants;
-import com.tricentis.common.reports.ExtentLogger;
+import com.tricentis.common.listeners.ITestListenerImp;
 import com.tricentis.common.reports.ExtentReport;
 import com.tricentis.common.utils.ConfigReader;
 import com.tricentis.web.drivers.DriverInstance;
 import com.tricentis.web.enums.DRIVERTYPE;
+import com.tricentis.web.reports.ExtentWebLogger;
 import org.apache.commons.io.FileUtils;
 import org.testng.ITestResult;
 import org.testng.annotations.*;
@@ -14,35 +15,45 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Objects;
 
+@Listeners(ITestListenerImp.class)
 public class BaseWebTest {
+
 
     @BeforeSuite
     public void beforeSuite() {
-        ExtentReport.setUpReport();
-        /*try {
-            //TODO handle empty file
-            FileUtils.cleanDirectory(new File(FrameworkConstants.SCREENSHOT_PATH));
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }*/
+
+        //ExtentReport.setUpReport();
+        File file = new File(FrameworkConstants.SCREENSHOT_PATH);
+        if (file.exists()) {
+            try {
+                FileUtils.cleanDirectory(new File(FrameworkConstants.SCREENSHOT_PATH));
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
     }
+
 
     @AfterSuite
     public void afterSuite() {
 
-        ExtentReport.closeReport();
+        //ExtentReport.closeReport();
     }
+
     @Parameters("browser")
     @BeforeMethod
     public void setUp(ITestResult result, @Optional String browser) {
-        if (Objects.isNull(browser)) {
-            browser = ConfigReader.getBrowserValue();
-        }
-        ExtentReport.createTestNode(result.getMethod().getMethodName());
-        DriverInstance.initDriver(DRIVERTYPE.valueOf(browser.toUpperCase()));
-        DriverInstance.getDriver().get(ConfigReader.getApplicationUrl());
-        DriverInstance.getDriver().manage().window().maximize();
-        ExtentLogger.info(ConfigReader.getApplicationUrl()+" URL is launched in "+browser+" in browser");
+
+            if (Objects.isNull(browser)) {
+                browser = ConfigReader.getBrowserValue();
+            }
+
+           ExtentReport.createTestNode(result.getMethod().getMethodName());
+            DriverInstance.initDriver(DRIVERTYPE.valueOf(browser.toUpperCase()));
+            DriverInstance.getDriver().get(ConfigReader.getApplicationUrl());
+            DriverInstance.getDriver().manage().window().maximize();
+            ExtentWebLogger.info(ConfigReader.getApplicationUrl() + " URL is launched in " + browser + " browser");
+
 
 
     }
@@ -50,14 +61,13 @@ public class BaseWebTest {
     @AfterMethod
     public void tearDown(ITestResult result) {
 
-        if(result.getStatus()==ITestResult.FAILURE){
-            ExtentLogger.fail("Failed because of - "+result.getThrowable().getMessage(),true);
-        }
 
         DriverInstance.quitDriver();
 
 
     }
+
+
 
 
 }
